+++ java-auth/src/main/java/com/example/app/service/CustomUserDetailsService.java
@@ -1,0 +1,38 @@
+package com.example.app.service;
+
+import com.example.app.domain.Member;
+import com.example.app.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.Collections;
+
+@Service
+@RequiredArgsConstructor
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return memberRepository.findByUsername(username)
+                .map(this::createUserDetails)
+                .orElseThrow(() -> new UsernameNotFoundException(username + " -> Not found in database."));
+    }
+
+    private UserDetails createUserDetails(Member member) {
+        GrantedAuthority grantedAuthority = new SimpleGrantedAuthority(member.getRole().name());
+
+        return new User(
+                member.getUsername(),
+                member.getPassword(),
+                Collections.singleton(grantedAuthority)
+        );
+    }
+}